// frontend/src/types/generated/financial/models/billingRuleCreate.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Conecta PRO - Financial Module
 * Gestão Financeira Completa - 251 endpoints em 15 submódulos
 * OpenAPI spec version: 1.0.0
 */
import type { BillingRuleCreateDescription } from './billingRuleCreateDescription';
import type { BillingType } from './billingType';
import type { BillingRuleCreateCategoryId } from './billingRuleCreateCategoryId';
import type { BillingRuleCreateBaseValue } from './billingRuleCreateBaseValue';
import type { BillingRuleCreateReferenceField } from './billingRuleCreateReferenceField';
import type { BillingFrequency } from './billingFrequency';
import type { BillingRuleCreateEndDate } from './billingRuleCreateEndDate';
import type { BillingRuleCreateInterestRate } from './billingRuleCreateInterestRate';
import type { BillingRuleCreatePenaltyRate } from './billingRuleCreatePenaltyRate';
import type { BillingRuleCreateDiscountRate } from './billingRuleCreateDiscountRate';
import type { BillingRuleCreateUnitFilter } from './billingRuleCreateUnitFilter';
import type { BillingRuleCreateNotes } from './billingRuleCreateNotes';

/**
 * Schema para criacao de regra.
 */
export interface BillingRuleCreate {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  description?: BillingRuleCreateDescription;
  billing_type?: BillingType;
  category_id?: BillingRuleCreateCategoryId;
  base_value: BillingRuleCreateBaseValue;
  value_type?: string;
  reference_field?: BillingRuleCreateReferenceField;
  frequency?: BillingFrequency;
  /**
   * @minimum 1
   * @maximum 28
   */
  due_day?: number;
  /**
   * @minimum 1
   * @maximum 28
   */
  generation_day?: number;
  start_date: string;
  end_date?: BillingRuleCreateEndDate;
  apply_interest?: boolean;
  interest_rate?: BillingRuleCreateInterestRate;
  apply_penalty?: boolean;
  penalty_rate?: BillingRuleCreatePenaltyRate;
  /** @minimum 0 */
  grace_days?: number;
  apply_discount?: boolean;
  discount_rate?: BillingRuleCreateDiscountRate;
  /** @minimum 0 */
  discount_days?: number;
  auto_generate_boleto?: boolean;
  /**
   * @minimum 1
   * @maximum 30
   */
  boleto_days_before?: number;
  /**
   * @minimum 1
   * @maximum 365
   */
  boleto_expiration_days?: number;
  auto_generate_pix?: boolean;
  /**
   * @minimum 1
   * @maximum 168
   */
  pix_expiration_hours?: number;
  notifications_enabled?: boolean;
  notification_channels?: string[];
  notify_before_days?: number[];
  notify_after_days?: number[];
  notification_time?: string;
  apply_to_all?: boolean;
  unit_filter?: BillingRuleCreateUnitFilter;
  notes?: BillingRuleCreateNotes;
  condominio_id: string;
}